[1699]: Move all the generic keeper test helper funcs to a new suite_test.go file as well as the TestSuite definition. Do a little cleanup to use those in a few places where it made sense.
