+++ x/exchange/keeper/genesis_test.go
@@ -1,177 +1,14 @@
 package keeper_test
 
 import (
-	"bytes"
 	"fmt"
-	"sort"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+
 	"github.com/provenance-io/provenance/x/exchange"
 	"github.com/provenance-io/provenance/x/exchange/keeper"
 )
-
-// copyGenState creates a copy of a GenesisState.
-func (s *TestSuite) copyGenState(genState *exchange.GenesisState) *exchange.GenesisState {
-	if genState == nil {
-		return nil
-	}
-	return &exchange.GenesisState{
-		Params:       s.copyParams(genState.Params),
-		Markets:      s.copyMarkets(genState.Markets),
-		Orders:       s.copyOrders(genState.Orders),
-		LastMarketId: genState.LastMarketId,
-		LastOrderId:  genState.LastOrderId,
-	}
-}
-
-// copyMarkets creates a copy of a slice of markets.
-func (s *TestSuite) copyMarkets(orig []exchange.Market) []exchange.Market {
-	return copySlice(orig, s.copyMarket)
-}
-
-// copyOrders creates a copy of a slice of orders.
-func (s *TestSuite) copyOrders(orig []exchange.Order) []exchange.Order {
-	return copySlice(orig, s.copyOrder)
-}
-
-// copyOrder creates a copy of an order.
-func (s *TestSuite) copyOrder(orig exchange.Order) exchange.Order {
-	rv := exchange.NewOrder(orig.OrderId)
-	switch {
-	case orig.IsAskOrder():
-		rv.WithAsk(s.copyAskOrder(orig.GetAskOrder()))
-	case orig.IsBidOrder():
-		rv.WithBid(s.copyBidOrder(orig.GetBidOrder()))
-	default:
-		rv.Order = orig.Order
-	}
-	return *rv
-}
-
-// copyAskOrder creates a copy of an AskOrder.
-func (s *TestSuite) copyAskOrder(orig *exchange.AskOrder) *exchange.AskOrder {
-	if orig == nil {
-		return nil
-	}
-	return &exchange.AskOrder{
-		MarketId:                orig.MarketId,
-		Seller:                  orig.Seller,
-		Assets:                  s.copyCoin(orig.Assets),
-		Price:                   s.copyCoin(orig.Price),
-		SellerSettlementFlatFee: s.copyCoinP(orig.SellerSettlementFlatFee),
-		AllowPartial:            orig.AllowPartial,
-		ExternalId:              orig.ExternalId,
-	}
-}
-
-// copyBidOrder creates a copy of a BidOrder.
-func (s *TestSuite) copyBidOrder(orig *exchange.BidOrder) *exchange.BidOrder {
-	if orig == nil {
-		return nil
-	}
-	return &exchange.BidOrder{
-		MarketId:            orig.MarketId,
-		Buyer:               orig.Buyer,
-		Assets:              s.copyCoin(orig.Assets),
-		Price:               s.copyCoin(orig.Price),
-		BuyerSettlementFees: s.copyCoins(orig.BuyerSettlementFees),
-		AllowPartial:        orig.AllowPartial,
-		ExternalId:          orig.ExternalId,
-	}
-}
-
-// copyParams creates a copy of exchange Params.
-func (s *TestSuite) copyParams(orig *exchange.Params) *exchange.Params {
-	if orig == nil {
-		return nil
-	}
-	return &exchange.Params{
-		DefaultSplit: orig.DefaultSplit,
-		DenomSplits:  s.copyDenomSplits(orig.DenomSplits),
-	}
-}
-
-// creates a copy of a DenomSplit.
-func (s *TestSuite) copyDenomSplit(orig exchange.DenomSplit) exchange.DenomSplit {
-	return exchange.DenomSplit{
-		Denom: orig.Denom,
-		Split: orig.Split,
-	}
-}
-
-// copyDenomSplits creates a copy of a slice of DenomSplits.
-func (s *TestSuite) copyDenomSplits(orig []exchange.DenomSplit) []exchange.DenomSplit {
-	return copySlice(orig, s.copyDenomSplit)
-}
-
-// sortGenState sorts the contents of a GenesisState.
-func (s *TestSuite) sortGenState(genState *exchange.GenesisState) *exchange.GenesisState {
-	if genState == nil {
-		return nil
-	}
-	if genState.Params != nil && len(genState.Params.DenomSplits) > 0 {
-		sort.Slice(genState.Params.DenomSplits, func(i, j int) bool {
-			return genState.Params.DenomSplits[i].Denom < genState.Params.DenomSplits[j].Denom
-		})
-	}
-	if len(genState.Markets) > 0 {
-		sort.Slice(genState.Markets, func(i, j int) bool {
-			return genState.Markets[i].MarketId < genState.Markets[j].MarketId
-		})
-		for _, market := range genState.Markets {
-			s.sortMarket(&market)
-		}
-	}
-	if len(genState.Orders) > 0 {
-		sort.Slice(genState.Orders, func(i, j int) bool {
-			return genState.Orders[i].OrderId < genState.Orders[j].OrderId
-		})
-	}
-	return genState
-}
-
-// sortMarket sorts all the fields in a market.
-func (s *TestSuite) sortMarket(market *exchange.Market) *exchange.Market {
-	if len(market.FeeSellerSettlementRatios) > 0 {
-		sort.Slice(market.FeeSellerSettlementRatios, func(i, j int) bool {
-			if market.FeeSellerSettlementRatios[i].Price.Denom < market.FeeSellerSettlementRatios[j].Price.Denom {
-				return true
-			}
-			if market.FeeSellerSettlementRatios[i].Price.Denom > market.FeeSellerSettlementRatios[j].Price.Denom {
-				return false
-			}
-			return market.FeeSellerSettlementRatios[i].Fee.Denom < market.FeeSellerSettlementRatios[j].Fee.Denom
-		})
-	}
-	if len(market.FeeBuyerSettlementRatios) > 0 {
-		sort.Slice(market.FeeBuyerSettlementRatios, func(i, j int) bool {
-			if market.FeeBuyerSettlementRatios[i].Price.Denom < market.FeeBuyerSettlementRatios[j].Price.Denom {
-				return true
-			}
-			if market.FeeBuyerSettlementRatios[i].Price.Denom > market.FeeBuyerSettlementRatios[j].Price.Denom {
-				return false
-			}
-			return market.FeeBuyerSettlementRatios[i].Fee.Denom < market.FeeBuyerSettlementRatios[j].Fee.Denom
-		})
-	}
-	if len(market.AccessGrants) > 0 {
-		sort.Slice(market.AccessGrants, func(i, j int) bool {
-			// Horribly inefficient. Not meant for production.
-			addrI, err := sdk.AccAddressFromBech32(market.AccessGrants[i].Address)
-			s.Require().NoError(err, "AccAddressFromBech32(%q)", market.AccessGrants[i].Address)
-			addrJ, err := sdk.AccAddressFromBech32(market.AccessGrants[j].Address)
-			s.Require().NoError(err, "AccAddressFromBech32(%q)", market.AccessGrants[j].Address)
-			return bytes.Compare(addrI, addrJ) < 0
-		})
-		for _, ag := range market.AccessGrants {
-			sort.Slice(ag.Permissions, func(i, j int) bool {
-				return ag.Permissions[i] < ag.Permissions[j]
-			})
-		}
-	}
-	return market
-}
 
 func (s *TestSuite) TestKeeper_InitAndExportGenesis() {
 	marketAcc := func(marketID uint32, name string) *exchange.MarketAccount {
